refactor(types): extract group tests to sibling file
